Fix pre-commit errors
