+++ tests/bgp/test_bgpmon.py
@@ -10,7 +10,7 @@ from tests.common.fixtures.ptfhost_utils import change_mac_addresses      # noqa
 from tests.common.fixtures.ptfhost_utils import remove_ip_addresses       # noqa F401
 from tests.common.helpers.generators import generate_ip_through_default_route
 from tests.common.helpers.assertions import pytest_assert
-from tests.common.utilities import wait_until, get_plt_reboot_ctrl
+from tests.common.utilities import wait_until
 from tests.common.utilities import wait_tcp_connection
 from bgp_helpers import BGPMON_TEMPLATE_FILE, BGPMON_CONFIG_FILE, BGP_MONITOR_NAME, BGP_MONITOR_PORT
 pytestmark = [
@@ -23,17 +23,6 @@ MAX_TIME_FOR_BGPMON = 180
 ZERO_ADDR = r'0.0.0.0/0'
 logger = logging.getLogger(__name__)
 
-
-@pytest.fixture(scope="module")
-def set_timeout_for_bgpmon(duthost):
-    """
-    For chassis testbeds, we need to specify plt_reboot_ctrl in inventory file,
-    to let MAX_TIME_TO_REBOOT to be overwritten by specified timeout value
-    """
-    global MAX_TIME_FOR_BGPMON
-    plt_reboot_ctrl = get_plt_reboot_ctrl(duthost, 'test_bgpmon.py', 'cold')
-    if plt_reboot_ctrl:
-        MAX_TIME_FOR_BGPMON = plt_reboot_ctrl.get('timeout', 180)
 
 def get_default_route_ports(host, tbinfo, default_addr=ZERO_ADDR):
     mg_facts = host.get_extended_minigraph_facts(tbinfo)
@@ -54,28 +43,6 @@ def get_default_route_ports(host, tbinfo, default_addr=ZERO_ADDR):
             port_indices.append(mg_facts['minigraph_ptf_indices'][port])
 
     return port_indices
-
-
-@pytest.fixture(scope="module")
-def dut_with_default_route(duthosts, enum_rand_one_per_hwsku_frontend_hostname, tbinfo):
-    if tbinfo['topo']['type'] == 't2':
-        # For T2 setup, default route via eBGP is only advertised from T3 VM's which are connected to one of the
-        # linecards and not the other. So, can't use enum_rand_one_per_hwsku_frontend_hostname for T2.
-        dut_to_T3 = None
-        for a_dut in duthosts.frontend_nodes:
-            minigraph_facts = a_dut.get_extended_minigraph_facts(tbinfo)
-            minigraph_neighbors = minigraph_facts['minigraph_neighbors']
-            for key, value in list(minigraph_neighbors.items()):
-                if 'T3' in value['name']:
-                    dut_to_T3 = a_dut
-                    break
-            if dut_to_T3:
-                break
-        if dut_to_T3 is None:
-            pytest.skip("Did not find any DUT in the DUTs (linecards) that are connected to T3 VM's")
-        return dut_to_T3
-    else:
-        return duthosts[enum_rand_one_per_hwsku_frontend_hostname]
 
 
 @pytest.fixture
